Removed unnecessary files from pages folder
