--- pages/blog/index.ts
@@ -1,3 +1,0 @@
-import Blog from './Blog';
-
-export default Blog;
