--- pages/login/index.ts
@@ -1,3 +1,0 @@
-import Login from './Login';
-
-export default Login;
