--- pages/meet/index.tsx
@@ -1,15 +1,0 @@
-import React, { ReactElement } from 'react';
-
-interface Props {
-
-}
-
-function Meet({ }: Props): ReactElement {
-	return (
-		<div>
-			Meet
-		</div>
-	);
-}
-
-export default Meet;
