--- pages/testimonials/index.ts
@@ -1,3 +1,0 @@
-import TestimonialsPage from './TestimonialsPage';
-
-export default TestimonialsPage;
